requirements.txt
Django==5.0.3
requests==2.28.2
DateTime==5.5
python-dateutil==2.8.2
numpy==1.25.2
opencv-python==4.9.0.80
Pillow==10.0.0